pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Code for the Agent modules: build your first Agent + Multi-Agent sytem"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crawl4ai>=0.6.1",
    "docling>=2.30.0",
    "duckduckgo-search>=8.0.1",
    "google-genai>=1.11.0",
    "pydantic-ai>=0.2.4",
    "pydantic-evals>=0.2.16",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.ruff]
line-length = 110
